+++ tmp-electron-sim/.next/standalone/.next/server/app/api/changelog/route_client-reference-manifest.js
@@ -1,0 +1,1 @@
+globalThis.__RSC_MANIFEST=(globalThis.__RSC_MANIFEST||{});globalThis.__RSC_MANIFEST["/api/changelog/route"]={"moduleLoading":{"prefix":"/_next/"},"ssrModuleMapping":{},"edgeSSRModuleMapping":{},"clientModules":{},"entryCSSFiles":{"D:\\Workspace\\Web\\Electron\\darkfloor-player\\apps\\web\\src\\app\\api\\changelog\\route":[]},"rscModuleMapping":{},"edgeRscModuleMapping":{}};